refactor(leaderboard): Separate data calculation into smaller util fns
